requirements.txt
requests==2.20.0
GitPython==2.1.11